pyproject.toml
[project]
name = "awscostintelligenceagent"
version = "0.1.0"
description = "A smart agent designed to analyze AWS usage and provide actionable cost-saving insights. Built using AWS native services (Lambda, Athena, Cost Explorer, CloudWatch), the agent delivers real-time cost breakdowns, forecasts, anomaly detection, and recommendations. Integrated with Slack/Teams for proactive alerts and custom reports for finance and engineering stakeholders"
authors = [
    {name = "om agrawal"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "openai (>=1.75.0,<2.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    
]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
langchain_community = "^0.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
